Update README.md
A lot of the steps are not obvious for a first time user of Twilio (and XCode), so I have added some more steps to make the process more obvious. Simple stuff like cloning the repo and and which XCode project to open just become simpler now. Hope this helps.

I'm still waiting for my Apple Developer creds, so will update past the Certificate part once there.